requirements.txt
spotipy>=2.24.0
spotipy-anon>=1.3.0
plexapi>=4.11.2
deezer-python>=2.3.0
spotdl>=4.2.8
python-dotenv>=1.0.1
pyyaml
colorama